Moved to another module called display_plot.py
